refactor(remediation): demote remediate_secret_incidents to data tool
Rename the doctrine-emitting remediate_secret_incidents tool to
list_remediation_targets and reframe it as a pure occurrence finder.
This removes the semantic collision with the agent-skills remediation
skill, which agents were reaching past in favor of the MCP tool.

- Rename tool + module + models (RemediateSecretIncidents* ->
  ListRemediationTargets*); result field remediation_instructions ->
  guidance, suggested_occurrences_for_remediation_count ->
  suggested_occurrences_count.
- Honest docstring/description: returns occurrence data only, does NOT
  rotate or modify files; defer to the remediation skill/workflow.
- Rewrite remediation_prompt.md as a rotation-first fallback that leads
  with rotate/revoke, gates git-history rewriting behind unpushed-only,
  and explicitly defers to a skill/workflow when present.
- Drop the now-unused git_commands/create_env_example/add_to_env params.
- Propagate rename through register_tools instructions, conftest,
  run_tool script, and tests; add a rotation-first guidance assertion.
